K3.12 fix for SGX modifying return value after gem_new
SGX modifies the value of buffer size after gem is created
to align it to page size. In 3.12 kernel, if gem allocation
size and vaddr start and end do not match, a mmap cannot
be done.

So why not determine how much will be the return size and
allocate a big enough gem area. Makes both parties happy.

Change-Id: I5fb99d7a479f53048ddb77bda001bf516bec5d06
Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>