K3.12 create a dummy gem handle for every gem allocation
K3.12 GEM vma features ensure that mmap on a gem is called
only if a handle is created for it. This serves the purpose
of authentication.

Since SGX creates GEM areas in KM using omap_gem_new_ext and
therefore no handles are created, mmap fails.

Lets just create a dummy handle for each gem object. Dont
bother to free it. drm_release will take care if it.

Change-Id: I2415a9a8aaf4001e558b824821a05904a2934005
Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>