km: Add more "SUPPORT_DRI_DRM_EXTERNAL" protected sections
The cuurent DDK supports only the OMAP5 devices which use external DRM driver
so that the SUPPORT_DRI_DRM_EXTERNAL is always defined. To enhnace the DDK to
support other types of devices which may not even use DRM, we need to enhance
the DDK KM to support the case where SUPPORT_DRI_DRM_EXTERNAL is not defined.

We need to add "#if defined(SUPPORT_DRI_DRM_EXTERNAL)" protection to any code
sections where omap_gen APIs or other omap_drm related structures or functionalities
are invoked.

Signed-off-by: Eric Ruei <e-ruei1@ti.com>