km: TEMP: map entire SGX address range
SGX needs entire register address range to be mapped. SoC integration only
requires GPU_WRAPPER range for hwmod.

Specify the GPU_WRAPPER range in DT. Map the entire range irrespective of what
is defined in DT.

The right way to do it would be through register definitions in DT that can
work for both hwmod as well as the driver.

This is a TEMP patch and should be addressed for DDK migration.

Signed-off-by: Anand Balagopalakrishnan <anandb@ti.com>