Makefile.template: lift the NOTINTERMEDIATE logic
Lift the NOTINTERMEDIATE logic from the kernel, as there is a feature
discrepancy in newer kbuild versions that causes build issues depending
on which version of kbuild is available for the kernel.

This is particularly an issue if you attempt to hijack the kernel module
build logic to build this module. If you are building it standalone you
wouldn't see any issues. In addition, this is only required for kernel
6.2 and up.

Thanks Masahiro Yamada <masahiroy@kernel.org> for the slick code.

Signed-off-by: Randolph Sapp <rs@ti.com>