config: core: use CONFIG_ZONE_DMA32 for LPAE
A previous patch switched this to only check for the DMA flag, not
DMA32. The point of these warnings was to indicate to the user that the
config they have selected has no safeguards. It's behavior is entirely
dependent on the hardware in use. Using just the presence of
CONFIG_ZONE_DMA completely masks this warning.

This reverts commit af9e23c3e85242a5607725ccb78840d1d0d3b2db.

Signed-off-by: Randolph Sapp <rs@ti.com>