compiler.mk: fix hardfloat toolchains not using suffix
This makes the build script check to see if a hardfloat value was
specified in SYS_CFLAGS for the given compiler before assuming whether
or not it supports hf based on it's suffix.

This properly addresses compilers, like the Yocto compilers, which
support hardfloat though they do not advertise it through the usual
channels (print-multi-lib or the compiler suffix).

This reverts commit 3882292ae3b787b09f35ea2853ddeb8b0e8f45df.

Signed-off-by: Randolph Sapp <rs@ti.com>